HOC take out as component
