+++ src/components/hoc-helper/index.js
@@ -1,0 +1,5 @@
+import withData from './withData'
+
+export {
+  withData
+};
